requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic-settings==2.1.0
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
alembic==1.13.1
greenlet==3.3.0
google-cloud-storage==2.14.0
reportlab==4.4.7